FIX use orf_num rather than index to flip reverse strand profiles
This is related to Issue #54.

Previously, the code assumed the orfs data frame was sorted in the same order
as the profiles. At one point, this was true; however, it seems the access
patterns and/or internal pandas data structures regarding indexes have changed
(see "What's new" for 0.19.2), and it is no longer correct. Thus, ORFs on the
reverse strand were not correctly flipped.

The code now explicitly uses orf_num to match orfs in the data frame with their
profile (which is the designed and expected behavior).