FIX use orf_num rather than index for filtering BF calculations
This is related to Issue #54, and it is essentially the same problem as fixed
in commit dieterich-lab/rp-bp@/0a9253d9aa53f4cccb187f1960cc3643d1fe87b0.

Again, the order of the orf data frame was assumed to match the order of the
orf profiles, and this assumption was no longer valid. The code now explicitly
uses orf_num rather than the index.