FIX removed subtract-one/add-one "fix" for starts
This addresses Issue #64.

It was unnecessary to make this adjustment to starts since bed intervals are
closed on that side. The current fix does implicitly imply that ORFs have a
genomic length of at least 1. Since the start codon alone is length 3, it is
not possible for a valid ORF to have length less than 3.

Test cases have been implemented which suggest the correctness of the approach,
and they will soon be incorporated into the package.